requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
redis>=5.0.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
djangorestframework-simplejwt>=5.0.0
Pillow>=10.0.0

django-extensions>=3.2.0
python-decouple>=3.8
celery>=5.3.0
django-celery-beat>=2.5.0
whitenoise>=6.6.0
gunicorn>=21.2.0 